Rename more things.
